+++ pricing/src/epilot/availability_api.py
@@ -27,7 +27,7 @@ class AvailabilityAPI:
         
         url = base_url + '/v1/public/availability:check'
         headers = utils.get_headers(request)
-        req_content_type, data, form = utils.serialize_request_body(request, "availability_check_params", False, False, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.DollarAvailabilityCheckRequest, "availability_check_params", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None: