+++ pricing/src/epilot/catalog_api.py
@@ -27,7 +27,7 @@ class CatalogAPI:
         
         url = base_url + '/v1/public/catalog'
         headers = utils.get_headers(request)
-        req_content_type, data, form = utils.serialize_request_body(request, "catalog_search", False, False, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.DollarSearchCatalogRequest, "catalog_search", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None: