+++ pricing/src/epilot/deprecated.py
@@ -31,7 +31,7 @@ class Deprecated:
         
         url = base_url + '/v1/public/opportunity'
         headers = utils.get_headers(request)
-        req_content_type, data, form = utils.serialize_request_body(request, "opportunity", False, False, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.DollarCreateOpportunityRequest, "opportunity", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None: