+++ pricing/src/epilot/models/errors/error.py
@@ -21,4 +21,4 @@ class Error(Exception):
     
 
     def __str__(self) -> str:
-        return utils.marshal_json(self)
+        return utils.marshal_json(self, type(self))