+++ pricing/src/epilot/order_api.py
@@ -25,7 +25,7 @@ class OrderAPI:
         
         url = base_url + '/v1/order'
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "request", False, False, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, shared.OrderPayload, "request", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None:
@@ -76,7 +76,7 @@ class OrderAPI:
         
         url = utils.generate_url(operations.PutOrderRequest, base_url, '/v1/order/{id}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "order_payload", False, False, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.PutOrderRequest, "order_payload", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None: